requirements.txt
loguru
yt_dlp
openai
requests
ffmpeg-python
faster-whisper
ctranslate2
numpy
hf_xet
tiktoken
pillow
b2sdk>=1.20.0